+++ integration/test_sen1_analysis.py
@@ -1,0 +1,44 @@
+#  Copyright (c) 2025-2026 by EOPF Sample Service team and contributors
+#  Permissions are hereby granted under the terms of the Apache 2.0 License:
+#  https://opensource.org/license/apache-2-0.
+
+from unittest import TestCase
+from pathlib import Path
+
+import xarray as xr
+
+from integration.helpers import assert_dataset_is_chunked
+from xarray_eopf.utils import timeit
+
+allowed_open_time = 1000  # seconds
+show_chunking = False
+
+
+class Sentinel2AnalysisTest(TestCase):
+    def test_open_dataset_sen1_grd(self):
+        dem_path = Path(__file__).resolve().parent / "test_data" / "dem_small.zarr"
+        dem = xr.open_dataset(str(dem_path), chunks={})
+        dem = dem.dem
+
+        url = (
+            "https://objects.eodc.eu/e05ab01a9d56408d82ac32d69a5aae2a:202603-"
+            "s01siwgrh-global/19/products/cpm_v262/S1A_IW_GRDH_1SDV_20260319"
+            "T102725_20260319T102758_063695_0801D3_2EC6.zarr"
+        )
+        with timeit("open " + url) as result:
+            # noinspection PyTypeChecker
+            ds = xr.open_dataset(
+                url,
+                engine="eopf-zarr",
+                op_mode="analysis",
+                dem=dem,
+                chunks={},
+            )
+        self.assertTrue(result.time_delta < allowed_open_time)
+
+        self.assertIn("vv", ds)
+        self.assertIn("vh", ds)
+
+        assert_dataset_is_chunked(self, ds, verbose=show_chunking)
+        for var_name in ds.data_vars:
+            self.assertEqual((541, 1081), ds[var_name].shape, msg=var_name)