+++ xarray_eopf/constants.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2025 by EOPF Sample Service team and contributors
+#  Copyright (c) 2025-2026 by EOPF Sample Service team and contributors
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
@@ -12,6 +12,7 @@ CONVERSION_FACTOR_DEG_METER = 111320
 
 FloatInt = float | int
 OpMode: TypeAlias = Literal["analysis", "native"]
+Sen1InterpMethods: TypeAlias = Literal["bilinear", "nearest"]
 
 # Keywords arguments passed to dataset.merge(other) when flattening
 # data trees.