+++ CHANGES.md
@@ -4,6 +4,7 @@
   (and its subpackages) is included in the PyPI wheel.
 - Remove the `coarsen.py` module, as it has been moved to [xcube-resampling](https://github.com/xcube-dev/xcube-resampling) 
   and is no longer used internally.
+- Add support for Sentinel-1 GRD analysis mode.
 
 
 ## Changes in 0.2.7 (from 2026-03-27)